requirements.txt
tavily-python
requests
numpy
rich
duckduckgo-search
requests
beautifulsoup4
python-dotenv